Add solution for c11pA and c11pB

+++ contest_6/problem_a/src/main.rs
@@ -61,7 +61,7 @@ impl LectureSleep {
     fn solve(self) -> Result<u32> {
         let technique_start_index = self
             .theorems
-            .windows(self.awake_duration as usize)
+            .windows(self.awake_duration)
             .enumerate()
             .map(|(index, theorems)| {
                 let theorems_by_technique = theorems
